Update user role selection and saving procedures
The way user roles are selected and saved has been updated. Rather than stringifying the whole roles array, only the first element is now appended to the form data in the UserAPI. This aligns with the change in the StoreEmployee file where only the first role from the token is used. The updateUser function has also been modified, instead of spreading the employee object, only necessary fields are explicitly stated.